Fixes part of #10415: Refactor email validators (#11077)
