+++ core/domain/prod_validators_test.py
@@ -24,7 +24,6 @@ import datetime
 import math
 import random
 import time
-import types
 
 from constants import constants
 from core.domain import collection_domain
@@ -2508,418 +2507,6 @@ class ConfigPropertySnapshotContentModelValidatorTests(
             u'[u\'fully-validated ConfigPropertySnapshotContentModel\', 2]']
         self.run_job_and_check_output(
             expected_output, sort=True, literal_eval=False)
-
-
-class SentEmailModelValidatorTests(test_utils.AuditJobsTestBase):
-
-    def setUp(self):
-        super(SentEmailModelValidatorTests, self).setUp()
-
-        def mock_generate_hash(
-                unused_cls, unused_recipient_id, unused_email_subject,
-                unused_email_body):
-            return 'Email Hash'
-
-        self.sender_email = 'noreply@oppia.org'
-        self.sender_id = 'sender'
-        self.sender_model = user_models.UserSettingsModel(
-            id=self.sender_id,
-            email=self.sender_email)
-        self.sender_model.update_timestamps()
-        self.sender_model.put()
-
-        self.recipient_email = 'recipient@email.com'
-        self.recipient_id = 'recipient'
-        self.recipient_model = user_models.UserSettingsModel(
-            id=self.recipient_id,
-            email=self.recipient_email)
-        self.recipient_model.update_timestamps()
-        self.recipient_model.put()
-
-        with self.swap(
-            email_models.SentEmailModel, '_generate_hash',
-            types.MethodType(mock_generate_hash, email_models.SentEmailModel)):
-            email_models.SentEmailModel.create(
-                self.recipient_id, self.recipient_email, self.sender_id,
-                self.sender_email, feconf.EMAIL_INTENT_SIGNUP,
-                'Email Subject', 'Email Body', datetime.datetime.utcnow())
-
-        self.model_instance = email_models.SentEmailModel.get_by_hash(
-            'Email Hash')[0]
-
-        self.job_class = (
-            prod_validation_jobs_one_off.SentEmailModelAuditOneOffJob)
-
-    def test_standard_model(self):
-        expected_output = [u'[u\'fully-validated SentEmailModel\', 1]']
-        self.run_job_and_check_output(
-            expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_created_on_greater_than_last_updated(self):
-        self.model_instance.created_on = (
-            self.model_instance.last_updated + datetime.timedelta(days=1))
-        self.model_instance.update_timestamps()
-        self.model_instance.put()
-        expected_output = [(
-            u'[u\'failed validation check for time field relation check '
-            'of SentEmailModel\', '
-            '[u\'Entity id %s: The created_on field has a value '
-            '%s which is greater than the value '
-            '%s of last_updated field\']]') % (
-                self.model_instance.id, self.model_instance.created_on,
-                self.model_instance.last_updated
-            )]
-        self.run_job_and_check_output(
-            expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_last_updated_greater_than_current_time(self):
-        self.model_instance.sent_datetime = (
-            datetime.datetime.utcnow() - datetime.timedelta(hours=20))
-        self.model_instance.update_timestamps()
-        self.model_instance.put()
-        expected_output = [(
-            u'[u\'failed validation check for current time check of '
-            'SentEmailModel\', '
-            '[u\'Entity id %s: The last_updated field has a '
-            'value %s which is greater than the time when the job was run\']]'
-        ) % (self.model_instance.id, self.model_instance.last_updated)]
-
-        mocked_datetime = datetime.datetime.utcnow() - datetime.timedelta(
-            hours=13)
-        with datastore_services.mock_datetime_for_datastore(mocked_datetime):
-            self.run_job_and_check_output(
-                expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_non_existent_sender_id(self):
-        self.sender_model.delete()
-        expected_output = [(
-            u'[u\'failed validation check for sender_id field check of '
-            'SentEmailModel\', '
-            '[u"Entity id %s: based on field sender_id having value '
-            '%s, expected model UserSettingsModel with '
-            'id %s but it doesn\'t exist"]]') % (
-                self.model_instance.id, self.sender_id, self.sender_id)]
-
-        self.run_job_and_check_output(
-            expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_non_existent_recipient_id(self):
-        self.recipient_model.delete()
-        expected_output = [(
-            u'[u\'failed validation check for recipient_id field check of '
-            'SentEmailModel\', '
-            '[u"Entity id %s: based on field recipient_id having value '
-            '%s, expected model UserSettingsModel with '
-            'id %s but it doesn\'t exist"]]') % (
-                self.model_instance.id, self.recipient_id, self.recipient_id)]
-
-        self.run_job_and_check_output(
-            expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_invalid_recipient_email(self):
-        self.recipient_model.email = 'invalid@email.com'
-        self.recipient_model.update_timestamps()
-        self.recipient_model.put()
-        expected_output = [(
-            u'[u\'failed validation check for recipient email check of '
-            'SentEmailModel\', '
-            '[u\'Entity id %s: Recipient email %s in entity does not match '
-            'with email %s of user obtained through recipient id %s\']]') % (
-                self.model_instance.id, self.model_instance.recipient_email,
-                self.recipient_model.email, self.model_instance.recipient_id)]
-        self.run_job_and_check_output(
-            expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_sent_datetime_greater_than_current_time(self):
-        self.model_instance.sent_datetime = (
-            datetime.datetime.utcnow() + datetime.timedelta(days=1))
-        self.model_instance.update_timestamps()
-        self.model_instance.put()
-        expected_output = [(
-            u'[u\'failed validation check for sent datetime check of '
-            'SentEmailModel\', '
-            '[u\'Entity id %s: The sent_datetime field has a value %s '
-            'which is greater than the time when the job was run\']]') % (
-                self.model_instance.id, self.model_instance.sent_datetime)]
-        self.run_job_and_check_output(
-            expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_invalid_id(self):
-        model_instance_with_invalid_id = email_models.SentEmailModel(
-            id='invalid', recipient_id=self.recipient_id,
-            recipient_email=self.recipient_email, sender_id=self.sender_id,
-            sender_email='noreply@oppia.org', intent=feconf.EMAIL_INTENT_SIGNUP,
-            subject='Email Subject', html_body='Email Body',
-            sent_datetime=datetime.datetime.utcnow())
-        model_instance_with_invalid_id.update_timestamps()
-        model_instance_with_invalid_id.put()
-        expected_output = [(
-            u'[u\'fully-validated SentEmailModel\', 1]'
-        ), (
-            u'[u\'failed validation check for model id check of '
-            'SentEmailModel\', '
-            '[u\'Entity id %s: Entity id does not match regex pattern\']]'
-        ) % 'invalid']
-        self.run_job_and_check_output(
-            expected_output, sort=True, literal_eval=False)
-
-
-class BulkEmailModelValidatorTests(test_utils.AuditJobsTestBase):
-
-    def setUp(self):
-        super(BulkEmailModelValidatorTests, self).setUp()
-
-        self.sender_email = 'sender@email.com'
-        self.sender_id = 'sender'
-        self.sender_model = user_models.UserSettingsModel(
-            id=self.sender_id,
-            email=self.sender_email)
-        self.sender_model.update_timestamps()
-        self.sender_model.put()
-
-        self.recipient_ids = ['recipient1', 'recipient2']
-        self.recipient_model_1 = user_models.UserSettingsModel(
-            id=self.recipient_ids[0],
-            email='recipient1@email.com')
-        self.recipient_model_1.update_timestamps()
-        self.recipient_model_1.put()
-        self.recipient_model_2 = user_models.UserSettingsModel(
-            id=self.recipient_ids[1],
-            email='recipient2@email.com')
-        self.recipient_model_2.update_timestamps()
-        self.recipient_model_2.put()
-
-        self.model_id = 'bulkemailid1'
-        email_models.BulkEmailModel.create(
-            self.model_id, self.recipient_ids, self.sender_id,
-            self.sender_email, feconf.BULK_EMAIL_INTENT_MARKETING,
-            'Email Subject', 'Email Body', datetime.datetime.utcnow())
-        self.model_instance = email_models.BulkEmailModel.get_by_id(
-            self.model_id)
-
-        self.job_class = (
-            prod_validation_jobs_one_off.BulkEmailModelAuditOneOffJob)
-
-    def test_standard_model(self):
-        expected_output = [u'[u\'fully-validated BulkEmailModel\', 1]']
-        self.run_job_and_check_output(
-            expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_created_on_greater_than_last_updated(self):
-        self.model_instance.created_on = (
-            self.model_instance.last_updated + datetime.timedelta(days=1))
-        self.model_instance.update_timestamps()
-        self.model_instance.put()
-        expected_output = [(
-            u'[u\'failed validation check for time field relation check '
-            'of BulkEmailModel\', '
-            '[u\'Entity id %s: The created_on field has a value '
-            '%s which is greater than the value '
-            '%s of last_updated field\']]') % (
-                self.model_instance.id, self.model_instance.created_on,
-                self.model_instance.last_updated
-            )]
-        self.run_job_and_check_output(
-            expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_last_updated_greater_than_current_time(self):
-        self.model_instance.sent_datetime = (
-            datetime.datetime.utcnow() - datetime.timedelta(hours=20))
-        self.model_instance.update_timestamps()
-        self.model_instance.put()
-        expected_output = [(
-            u'[u\'failed validation check for current time check of '
-            'BulkEmailModel\', '
-            '[u\'Entity id %s: The last_updated field has a '
-            'value %s which is greater than the time when the job was run\']]'
-        ) % (self.model_instance.id, self.model_instance.last_updated)]
-
-        mocked_datetime = datetime.datetime.utcnow() - datetime.timedelta(
-            hours=13)
-        with datastore_services.mock_datetime_for_datastore(mocked_datetime):
-            self.run_job_and_check_output(
-                expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_non_existent_sender_id(self):
-        self.sender_model.delete()
-        expected_output = [(
-            u'[u\'failed validation check for sender_id field check of '
-            'BulkEmailModel\', '
-            '[u"Entity id %s: based on field sender_id having value '
-            '%s, expected model UserSettingsModel with '
-            'id %s but it doesn\'t exist"]]') % (
-                self.model_instance.id, self.sender_id, self.sender_id)]
-
-        self.run_job_and_check_output(
-            expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_non_existent_recipient_id(self):
-        self.recipient_model_1.delete()
-        expected_output = [(
-            u'[u\'failed validation check for recipient_id field check of '
-            'BulkEmailModel\', '
-            '[u"Entity id %s: based on field recipient_id having value '
-            '%s, expected model UserSettingsModel with '
-            'id %s but it doesn\'t exist"]]') % (
-                self.model_instance.id, self.recipient_ids[0],
-                self.recipient_ids[0])]
-
-        self.run_job_and_check_output(
-            expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_invalid_sender_email(self):
-        self.sender_model.email = 'invalid@email.com'
-        self.sender_model.update_timestamps()
-        self.sender_model.put()
-        expected_output = [(
-            u'[u\'failed validation check for sender email check of '
-            'BulkEmailModel\', '
-            '[u\'Entity id %s: Sender email %s in entity does not match with '
-            'email %s of user obtained through sender id %s\']]') % (
-                self.model_instance.id, self.model_instance.sender_email,
-                self.sender_model.email, self.model_instance.sender_id)]
-        self.run_job_and_check_output(
-            expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_sent_datetime_greater_than_current_time(self):
-        self.model_instance.sent_datetime = (
-            datetime.datetime.utcnow() + datetime.timedelta(days=1))
-        self.model_instance.update_timestamps()
-        self.model_instance.put()
-        expected_output = [(
-            u'[u\'failed validation check for sent datetime check of '
-            'BulkEmailModel\', '
-            '[u\'Entity id %s: The sent_datetime field has a value %s '
-            'which is greater than the time when the job was run\']]') % (
-                self.model_instance.id, self.model_instance.sent_datetime)]
-        self.run_job_and_check_output(
-            expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_invalid_id(self):
-        model_instance_with_invalid_id = email_models.BulkEmailModel(
-            id='invalid:id', recipient_ids=self.recipient_ids,
-            sender_id=self.sender_id, sender_email=self.sender_email,
-            intent=feconf.BULK_EMAIL_INTENT_MARKETING,
-            subject='Email Subject', html_body='Email Body',
-            sent_datetime=datetime.datetime.utcnow())
-        model_instance_with_invalid_id.update_timestamps()
-        model_instance_with_invalid_id.put()
-        expected_output = [(
-            u'[u\'fully-validated BulkEmailModel\', 1]'
-        ), (
-            u'[u\'failed validation check for model id check of '
-            'BulkEmailModel\', '
-            '[u\'Entity id %s: Entity id does not match regex pattern\']]'
-        ) % model_instance_with_invalid_id.id]
-        self.run_job_and_check_output(
-            expected_output, sort=True, literal_eval=False)
-
-
-class GeneralFeedbackEmailReplyToIdModelValidatorTests(
-        test_utils.AuditJobsTestBase):
-
-    def setUp(self):
-        super(GeneralFeedbackEmailReplyToIdModelValidatorTests, self).setUp()
-
-        self.thread_id = feedback_services.create_thread(
-            'exploration', 'expid', None, 'a subject', 'some text')
-
-        self.signup(USER_EMAIL, USER_NAME)
-        self.user_id = self.get_user_id_from_email(USER_EMAIL)
-
-        self.model_instance = (
-            email_models.GeneralFeedbackEmailReplyToIdModel.create(
-                self.user_id, self.thread_id))
-        self.model_instance.update_timestamps()
-        self.model_instance.put()
-
-        self.job_class = (
-            prod_validation_jobs_one_off
-            .GeneralFeedbackEmailReplyToIdModelAuditOneOffJob)
-
-    def test_standard_model(self):
-        expected_output = [(
-            u'[u\'fully-validated GeneralFeedbackEmailReplyToIdModel\', 1]')]
-        self.run_job_and_check_output(
-            expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_created_on_greater_than_last_updated(self):
-        self.model_instance.created_on = (
-            self.model_instance.last_updated + datetime.timedelta(days=1))
-        self.model_instance.update_timestamps()
-        self.model_instance.put()
-        expected_output = [(
-            u'[u\'failed validation check for time field relation check '
-            'of GeneralFeedbackEmailReplyToIdModel\', '
-            '[u\'Entity id %s: The created_on field has a value '
-            '%s which is greater than the value '
-            '%s of last_updated field\']]') % (
-                self.model_instance.id, self.model_instance.created_on,
-                self.model_instance.last_updated
-            )]
-        self.run_job_and_check_output(
-            expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_last_updated_greater_than_current_time(self):
-        expected_output = [(
-            u'[u\'failed validation check for current time check of '
-            'GeneralFeedbackEmailReplyToIdModel\', '
-            '[u\'Entity id %s: The last_updated field has a '
-            'value %s which is greater than the time when the job was run\']]'
-        ) % (self.model_instance.id, self.model_instance.last_updated)]
-
-        mocked_datetime = datetime.datetime.utcnow() - datetime.timedelta(
-            hours=13)
-        with datastore_services.mock_datetime_for_datastore(mocked_datetime):
-            self.run_job_and_check_output(
-                expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_non_existent_user_id(self):
-        user_models.UserSettingsModel.get_by_id(self.user_id).delete()
-        expected_output = [(
-            u'[u\'failed validation check for item.id.user_id field check of '
-            'GeneralFeedbackEmailReplyToIdModel\', '
-            '[u"Entity id %s: based on field item.id.user_id having value '
-            '%s, expected model UserSettingsModel with '
-            'id %s but it doesn\'t exist"]]') % (
-                self.model_instance.id, self.user_id, self.user_id)]
-
-        self.run_job_and_check_output(
-            expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_non_existent_thread_id(self):
-        feedback_models.GeneralFeedbackThreadModel.get_by_id(
-            self.thread_id).delete()
-        expected_output = [(
-            u'[u\'failed validation check for item.id.thread_id field check of '
-            'GeneralFeedbackEmailReplyToIdModel\', '
-            '[u"Entity id %s: based on field item.id.thread_id having value '
-            '%s, expected model GeneralFeedbackThreadModel with '
-            'id %s but it doesn\'t exist"]]') % (
-                self.model_instance.id, self.thread_id, self.thread_id)]
-
-        self.run_job_and_check_output(
-            expected_output, sort=False, literal_eval=False)
-
-    def test_model_with_invalid_reply_to_id(self):
-        while len(
-                self.model_instance.reply_to_id) <= (
-                    email_models.REPLY_TO_ID_LENGTH):
-            self.model_instance.reply_to_id = (
-                self.model_instance.reply_to_id + 'invalid')
-        self.model_instance.update_timestamps()
-        self.model_instance.put()
-        expected_output = [(
-            u'[u\'failed validation check for reply_to_id length check of '
-            'GeneralFeedbackEmailReplyToIdModel\', '
-            '[u\'Entity id %s: reply_to_id %s should have length less than or '
-            'equal to %s but instead has length %s\']]'
-        ) % (
-            self.model_instance.id, self.model_instance.reply_to_id,
-            email_models.REPLY_TO_ID_LENGTH,
-            len(self.model_instance.reply_to_id))]
-        self.run_job_and_check_output(
-            expected_output, sort=False, literal_eval=False)
 
 
 class ExplorationModelValidatorTests(test_utils.AuditJobsTestBase):
